+++ resources/services.php
@@ -27,6 +27,7 @@ use Symfony\Lsp\Feature\DocumentLinkProviderRegistry;
 use Symfony\Lsp\Feature\Environment\EnvironmentCompletionProvider;
 use Symfony\Lsp\Feature\Environment\EnvironmentDiagnosticProvider;
 use Symfony\Lsp\Feature\Environment\EnvironmentRelationshipProvider;
+use Symfony\Lsp\Feature\EnvironmentScopedDiagnosticFilter;
 use Symfony\Lsp\Feature\Event\EventCodeLensProvider;
 use Symfony\Lsp\Feature\Event\EventSubscriberMapAnalyzer;
 use Symfony\Lsp\Feature\Event\EventYamlListenerAnalyzer;
@@ -145,6 +146,7 @@ return static function (ContainerConfigurator $container): void {
     $services->set(DiagnosticCollector::class);
     $services->set(DiagnosticSuppressor::class);
     $services->set(PartialParseDiagnosticFilter::class);
+    $services->set(EnvironmentScopedDiagnosticFilter::class);
     $services->set(UnknownNameCodeActionBuilder::class);
     $services->set(RenameEditBuilder::class);
     $featureGroups = [